refactor: create bootstrap package (PR1)
- Create handoff/bootstrap/ with paths, config, logging, docs
- Move paths.py, config.py, logging.py, docs.py into bootstrap/
- Update get_app_root() for new bootstrap path (parents[3])
- Update all imports: updater, update_ui, ui, pages, data/io,
  services/settings_service, __main__, scripts/cli, tests
- Delete old top-level module files
