+++ src/handoff/bootstrap/paths.py
@@ -7,4 +7,4 @@ from pathlib import Path
 
 def get_app_root() -> Path:
     """Return the root directory where ``app.py`` and ``src/`` live."""
-    return Path(__file__).resolve().parents[2]
+    return Path(__file__).resolve().parents[3]
