+++ src/handoff/data/io.py
@@ -7,8 +7,8 @@ from typing import Any
 from sqlmodel import select
 
 from handoff.backup_schema import BackupPayload
+from handoff.bootstrap.logging import log_application_action
 from handoff.db import session_context
-from handoff.logging import log_application_action
 from handoff.models import CheckIn, Handoff, Project
