+++ src/handoff/services/settings_service.py
@@ -189,6 +189,6 @@ def import_payload(payload: dict[str, Any]) -> None:
 
 
 def log_application_action(action: str, **details: Any) -> None:
-    import handoff.logging as _logging
+    import handoff.bootstrap.logging as _logging
 
     _logging.log_application_action(action, **details)
