+++ tests/test_config.py
@@ -1,4 +1,4 @@
-"""Tests for handoff.config (Streamlit env defaults).
+"""Tests for handoff.bootstrap.config (Streamlit env defaults).
 
 Config is applied at import time. We run assertions in a subprocess so the
 test runner's environment is not mutated and tests remain order-independent.
@@ -15,7 +15,7 @@ PROJECT_ROOT = Path(__file__).resolve().parents[1]
 
 
 def test_config_sets_streamlit_env_defaults() -> None:
-    """Importing handoff.config sets STREAMLIT_* env vars when not already set."""
+    """Importing handoff.bootstrap.config sets STREAMLIT_* env vars when not already set."""
     code = """
 import os
 # Simulate clean env for the keys we set (so setdefault takes effect).
@@ -27,7 +27,7 @@ for key in (
     "STREAMLIT_BROWSER_GATHER_USAGE_STATS",
 ):
     os.environ.pop(key, None)
-import handoff.config  # noqa: F401
+import handoff.bootstrap.config  # noqa: F401
 assert os.environ.get("STREAMLIT_CLIENT_SHOW_ERROR_DETAILS") == "none"
 assert os.environ.get("STREAMLIT_CLIENT_TOOLBAR_MODE") == "viewer"
 assert os.environ.get("STREAMLIT_CLIENT_SHOW_SIDEBAR_NAVIGATION") == "false"
